+++ fast64_internal/oot/animation/exporter/__init__.py
@@ -1,0 +1,4 @@
+from .functions import (
+    ootExportLinkAnimation,
+    ootExportNonLinkAnimation,
+)
