+++ fast64_internal/oot/animation/importer/__init__.py
@@ -1,0 +1,4 @@
+from .functions import (
+    ootImportLinkAnimationC,
+    ootImportNonLinkAnimationC,
+)
