+++ fast64_internal/oot/animation/operators.py
@@ -5,13 +5,8 @@ from bpy.utils import register_class, unregister_class
 from bpy.ops import object
 from ...utility import PluginError, toAlnum, writeCData, raisePluginError
 from .properties import OOTAnimExportSettingsProperty, OOTAnimImportSettingsProperty
-
-from ..oot_anim import (
-    ootExportLinkAnimation,
-    ootExportNonLinkAnimation,
-    ootImportLinkAnimationC,
-    ootImportNonLinkAnimationC,
-)
+from .exporter import ootExportLinkAnimation, ootExportNonLinkAnimation
+from .importer import ootImportLinkAnimationC, ootImportNonLinkAnimationC
 
 from ..oot_utility import (
     ootGetPath,
